requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.12.3
pydantic-settings==2.6.1
python-dotenv==1.1.1
openai==1.44.0
httpx==0.27.2
httpcore==1.0.5
faiss-cpu==1.12.0
numpy>=1.26,<3.0
pypdf==5.0.1
python-docx==1.1.2
markdown==3.7

# NEW (OCR + PDF rendering)
pytesseract==0.3.13
Pillow==10.4.0
pypdfium2==4.30.0
